완성

+++ routes/gbook-router.js
@@ -32,29 +32,44 @@ router.get(['/', '/list', '/list/:page'], async (req, res, next) => {
 	}
 });
 
-router.post('/create', isUser, upload.single('upfile'), async (req, res, next) => {
+router.post('/save', isUser, upload.single('upfile'), async (req, res, next) => {
 	try {
 		let sql, values; 
-
-		// gbook 저장
-		let { writer, content } = req.body;
-		sql = 'INSERT INTO gbook SET writer=?, content=?, uid=?';
-		values = [writer, content, req.session.user.id];
-		const [r] = await pool.execute(sql, values);
-
-		if(req.file) {
-			//gbookfile 저장
-			let { originalname, filename, size, mimetype } = req.file;
-			sql = 'INSERT INTO gbookfile SET oriname=?, savename=?, size=?, type=?, gid=?';
-			values = [originalname, filename, size, mimetype, r.insertId];
-			const [r2] = await pool.execute(sql, values);
+		let { writer, content, id } = req.body;
+		if(id && id !== '') { // 수정
+			sql = 'UPDATE gbook SET writer=?, content=? WHERE id=? AND uid=?';
+			var [r] = await pool.execute(sql, [writer, content, id, req.session.user.id]);
+			r.id = id;
 		}
-		res.send(alert('저장되었습니다', '/gbook'));
+		else { // 저장
+			sql = 'INSERT INTO gbook SET writer=?, content=?, uid=?';
+			var [r] = await pool.execute(sql, [writer, content, req.session.user.id]);
+			r.id = r.insertId;
+		}
+		if(req.file) { // 첨부파일 처리
+			let { originalname, filename, size, mimetype, isExist = false } = req.file;
+			if(id && id !== '') { // 수정에서 기존 파일 삭제
+				sql = 'SELECT savename FROM gbookfile WHERE gid=?';
+				const [r2] = await pool.execute(sql, [id]);
+				if(r2.length) {
+					await fs.remove(transBackSrc(r2[0].savename));
+					isExist = true;
+				}
+			}
+			sql = isExist ? 
+				'UPDATE gbookfile SET oriname=?, savename=?, size=?, type=? WHERE gid=?' :
+				'INSERT INTO gbookfile SET oriname=?, savename=?, size=?, type=?, gid=?' ;
+			values = [originalname, filename, size, mimetype, r.id];
+			const [r3] = await pool.execute(sql, values); // gbookfile 처리
+		}
+		if(id && id !== '') res.send(alert('수정되었습니다.', '/'));
+		else res.send(alert('저장되었습니다.', '/'));
 	}
 	catch(err) {
 		next(error(err));
 	}
 });
+
 
 router.get('/remove/:id', isUser, async (req, res, next) => {
 	try {
@@ -65,7 +80,7 @@ router.get('/remove/:id', isUser, async (req, res, next) => {
 		sql = 'DELETE FROM gbook WHERE id=? AND uid=?'; // 글 레코드 삭제함 -> 첨부파일 레코드도 삭제됨
 		const [r2] = await pool.execute(sql, [id, req.session.user.id]);
 		if(r2.affectedRows === 1) { // 글 레코드 및 첨부파일 레코드가 삭제됐다면...
-			await fs.remove(transBackSrc(r[0].savename));	// 실제 첨부파일 삭제
+			if(r.length === 1 ) await fs.remove(transBackSrc(r[0].savename));	// 실제 첨부파일 삭제
 			res.redirect('/');
 		}
 		else {
@@ -77,19 +92,48 @@ router.get('/remove/:id', isUser, async (req, res, next) => {
 	}
 });
 
-router.get('/view/:id', isUser, async(req, res, next) => { //마지막에 위치 (주소값을 아이디값으로 인식할 수 있기때문에 )
-	try{
+router.get('/view/:id', isUser, async (req, res, next) => {
+	try {
 		let sql, values;
-		sql = 'SELECT G.*, F.savename FROM gbook G LEFT JOIN gbookfile F ON G.id = F.gid WHERE G.id=? AND G.uid=?';
+		sql = 'SELECT G.*, F.savename, F.id AS fid FROM gbook G LEFT JOIN gbookfile F ON G.id = F.gid WHERE G.id=? AND G.uid=?';
 		const [r] = await pool.execute(sql, [req.params.id, req.session.user.id]);
-		if(r.length == 1){
+		if(r.length == 1) {
 			r[0].savename = transFrontSrc(r[0].savename);
-			res.status(200).json({ code: 200, success: true, data: r[0]});
-
-		}else res.status(200).json({ code: 200, success: false, data: null }); //통신 성공 데이타를 못가져옴
-	}catch(err){
-		res.status(500).json({ code: 500, err }); 
+			res.status(200).json({ code: 200, success: true, data: r[0] });
+		}
+		else {
+			res.status(200).json({ code: 200, success: false, data: null });
+		}
 	}
-}); 
+	catch(err) {
+		res.status(500).json({ code: 500, err });
+	}
+});
+
+router.get('/file/remove', isUser, async (req, res, next) => {
+	try {
+		let sql, values;
+		sql = 'SELECT F.savename FROM gbook G LEFT JOIN gbookfile F ON G.id = F.gid WHERE G.id=? AND G.uid=? AND F.id=?';
+		const [r] = await pool.execute(sql, [req.query.id, req.session.user.id, req.query.fid]);
+		if(r.length === 1) {
+			sql = 'DELETE FROM gbookfile WHERE id=?';
+			const [r2] = await pool.execute(sql, [req.query.fid]);
+			if(r2.affectedRows === 1) {
+				await fs.remove(transBackSrc(r[0].savename));
+				res.status(200).json({ code: 200, success: true });
+			}
+			else {
+				res.status(200).json({ code: 200, success: false });
+			}
+		}
+		else {
+			res.status(200).json({ code: 200, success: false });
+		}
+	}
+	catch(err) {
+		console.log(err);
+		res.status(500).json({ code: 500, err });
+	}
+});
 
 module.exports = router;
